+++ R/fit_drc_4p.R
@@ -1,41 +1,49 @@
 #' Fitting four-parameter dose response curves
 #'
 #' Function for fitting four-parameter dose response curves for each group (precursor, peptide or
-#' protein). In addition it can filter data based on completeness, the completeness distribution
-#' and statistical testing using ANOVA.
+#' protein). In addition it can annotate data based on completeness, the completeness distribution
+#' and statistical testing using ANOVA. Filtering by the function is only performed based on completeness
+#' if selected.
 #'
-#' @details If data filtering options are selected, data is filtered based on multiple criteria.
-#' In general, curves are only fitted if there are at least 5 conditions with data points present
-#' to ensure that there is potential for a good curve fit. Therefore, this is also the case if no
-#' filtering option is selected. Furthermore, a completeness cutoff is defined for filtering. By
-#' default each entity (e.g. precursor) is filtered to contain at least 70% of total replicates
-#' (adjusted downward) for at least 50% of all conditions (adjusted downward). This can be adjusted
-#' with the according arguments. In addition to the completeness cutoff, also a significance cutoff
-#' is applied. ANOVA is used to compute the statistical significance of the change for each entity.
-#' The resulting p-value is adjusted using the Benjamini-Hochberg method and a cutoff of q <= 0.05
-#' is applied. Curve fits that have a minimal value that is higher than the maximal value are
-#' excluded as they were likely wrongly fitted. Curves with a correlation below 0.8 are not passing
-#' the filtering. If a fit does not fulfill the significance or completeness cutoff, it has a chance
-#' to still be considered if half of its values (+/-1 value) pass the replicate completeness
-#' criteria and half do not pass it. In order to fall into this category, the values that fulfill t
-#' he completeness cutoff and the ones that do not fulfill it need to be consecutive, meaning
-#' located next to each other based on their concentration values. Furthermore, the values that
-#' do not pass the completeness cutoff need to be lower in intensity. Lastly, the difference
+#' @details If data filtering options are selected, data is annotated based on multiple criteria.
+#' If `"post"` is selected the data is annotated based on completeness, the completeness distribution, the
+#' adjusted ANOVA p-value cutoff and a correlation cutoff. Completeness of features is determined based on
+#' the `n_replicate_completeness` and `n_condition_completeness` arguments. The completeness distribution determines
+#' if there is a distribution of not random missingness of data along the dose. For this it is checked if half of a
+#' features values (+/-1 value) pass the replicate completeness criteria and half do not pass it. In order to fall into
+#' this category, the values that fulfill the completeness cutoff and the ones that do not fulfill it
+#' need to be consecutive, meaning located next to each other based on their concentration values. Furthermore,
+#' the values that do not pass the completeness cutoff need to be lower in intensity. Lastly, the difference
 #' between the two groups is tested for statistical significance using a Welch's t-test and a
-#' cutoff of p <= 0.1 (we want to mainly discard curves that falsly fit the other criteria but that
+#' cutoff of p <= 0.1 (we want to mainly discard curves that falsely fit the other criteria but that
 #' have clearly non-significant differences in mean). This allows curves to be considered that have
 #' missing values in half of their observations due to a decrease in intensity. It can be thought
 #' of as conditions that are missing not at random (MNAR). It is often the case that those entities
 #' do not have a significant p-value since half of their conditions are not considered due to data
-#' missingness.
+#' missingness. The ANOVA test is performed on the features by concentration. If it is significant it is
+#' likely that there is some response. However, this test would also be significant even if there is one
+#' outlier concentration so it should only be used only in combination with other cutoffs to determine
+#' if a feature is significant. The `passed_filter` column is `TRUE` for all the
+#' features that pass the above mentioned criteria and that have a correlation greater than the cutoff
+#' (default is 0.8) and the adjusted ANOVA p-value below the cutoff (default is 0.05).
 #'
-#' The final filtered list is ranked based on a score calculated on entities that pass the filter.
+#' The final list is ranked based on a score calculated on entities that pass the filter.
 #' The score is the negative log10 of the adjusted ANOVA p-value scaled between 0 and 1 and the
 #' correlation scaled between 0 and 1 summed up and divided by 2. Thus, the highest score an
 #' entity can have is 1 with both the highest correlation and adjusted p-value. The rank is
 #' corresponding to this score. Please note, that entities with MNAR conditions might have a
-#' lower score due to the missing or non-significant ANOVA p-value. You should have a look at
-#' curves that are TRUE for \code{dose_MNAR} in more detail.
+#' lower score due to the missing or non-significant ANOVA p-value. If no score could be calculated
+#' the usual way these cases receive a score of 0. You should have a look at curves that are TRUE
+#' for \code{dose_MNAR} in more detail.
+#'
+#' If the `"pre"` option is selected for the `filter` argument then the data is filtered for completeness
+#' prior to curve fitting and the ANOVA test. Otherwise annotation is performed exactly as mentioned above.
+#' We recommend the `"pre"` option because it leaves you with not only the likely hits of your treatment, but
+#' also with rather high confidence true negative results. This is because the filtered data has a high
+#' degree of completeness making it unlikely that a real dose-response curve is missed due to data missingness.
+#'
+#' Please note that in general, curves are only fitted if there are at least 5 conditions with data points present
+#' to ensure that there is potential for a good curve fit. This is done independent of the selected filtering option.
 #'
 #' @param data a data frame that contains at least the input variables.
 #' @param sample  a character column in the \code{data} data frame that contains the sample names.
@@ -45,18 +53,17 @@
 #' values, e.g. log2 transformed intensities.
 #' @param dose  a numeric column in the \code{data} data frame that contains the dose values, e.g.
 #' the treatment concentrations.
-#' @param filter a character value that determines if models should be filtered and if they should
-#' be filtered before or after the curve fits. Filtering of models can be skipped with
-#' \code{filter = "none"}. Data can be filtered prior to model fitting with \code{filter = "pre"}.
-#' In that case models will only be fitted for data that passed the filtering step. This will
-#' allow for faster model fitting since only fewer models will be fit. If you plan on performing
-#' an enrichment analysis you have to choose \code{filter = "post"}. All models will be fit (even
-#' the ones that do not pass the filtering criteria). For enrichment analysis you should use both
-#' good (i.e. models that pass the filtering) and bad (i.e. models that do not pass the filtering)
-#' models. Therefore, for post-filtering the full list is returned and it will only contain
-#' annotations that indicate (\code{passed_filter}) if the filtering was passed or not. Default is
-#' "post". For ANOVA an adjusted p-value of 0.05 is used as a cutoff.
-#' @param replicate_completeness a numeric value which similar to \code{completenss_MAR} of the
+#' @param filter a character value that can either be `"pre"`, `"post"` or `"none"`. The data is
+#' annotated for completeness, ANOVA significance and the completeness distribution along
+#' the doses (`"pre"` and `"post"`). The combined output of this filtering step can be found in
+#' the `passed_filter` column and depends on the cutoffs provided to the function. Note that this
+#' is only an annotation and nothing is removed from the output. If `"pre"` is selected then, in
+#' addition to the annotation, the data is filtered for completeness based on the condition completeness
+#' prior to the curve fitting and ANOVA calculation and p-value adjustment. This has the benefit that less
+#' curves need to be fitted and that the ANOVA p-value adjustment is done only on the relevant set of tests.
+#' If `"none"` is selected the data will be neither annotated nor filtered.
+#' @param replicate_completeness `r lifecycle::badge("deprecated")` please use `n_replicate_completeness` instead.
+#' A numeric value which similar to \code{completenss_MAR} of the
 #' \code{assign_missingness} function sets a threshold for the completeness of data. In contrast
 #' to \code{assign_missingness} it only determines the completeness for one condition and not the
 #' comparison of two conditions. The threshold is used to calculate a minimal degree of data
@@ -64,14 +71,31 @@
 #' It is multiplied with the number of replicates and then adjusted downward. The resulting number
 #' is the minimal number of observations that a condition needs to have to be considered "complete
 #' enough" for the \code{condition_completeness} argument.
-#' @param condition_completeness a numeric value which determines how many conditions need to at
+#' @param condition_completeness `r lifecycle::badge("deprecated")` please use `n_condition_completeness` instead.
+#' A numeric value which determines how many conditions need to at
 #' least fulfill the "complete enough" criteria set with \code{replicate_completeness}. The
 #' value provided to this argument has to be between 0 and 1, default is 0.5. It is multiplied with
 #' the number of conditions and then adjusted downward. The resulting number is the minimal number
 #' of conditions that need to fulfill the \code{replicate_completeness} argument for a peptide to
 #' pass the filtering.
-#' @param correlation_cutoff a numeric vector that specifies the correlation cutoff used for data
-#' filtering.
+#' @param n_replicate_completeness a numeric value that defines the minimal number of observations that a
+#' condition (concentration) needs to have to be considered "complete enough" for the `n_condition_completeness`
+#' argument. E.g. if each concentration has 4 replicates this argument could be set to 3 to allow for one
+#' replicate to be missing for the completeness criteria.
+#' @param n_condition_completeness a numeric value that defines the minimal number
+#' of conditions that need to fulfill the `n_replicate_completeness` argument for a feature to
+#' pass the filtering. E.g. if an experiment has 12 concentrations, this argument could be set
+#' to 6 to define that at least 6 of 12 concentrations need to make the replicate completeness cutoff.
+#' @param complete_doses an optional numeric vector that supplies all the actually used doses (concentrations)
+#' to the function. Usually the function extracts this information from the supplied data. However,
+#' for incomplete datasets the total number of assumed doses might be wrong. Therefore, it becomes
+#' important to provide this argument when the dataset is small and potentially incomplete. This
+#' information is only used for the missing not at random (MNAR) estimations.
+#' @param anova_cutoff a numeric value that specifies the ANOVA adjusted p-value cutoff used for
+#' data filtering. Any fits with an adjusted ANOVA p-value bellow the cutoff will be considered
+#' for scoring. The default is `0.05`.
+#' @param correlation_cutoff a numeric value that specifies the correlation cutoff used for data
+#' filtering. Any fits with a correlation above the cutoff will be considered for scoring.
 #' @param log_logarithmic a logical value that indicates if a logarithmic or log-logarithmic model
 #' is fitted. If response values form a symmetric curve for non-log transformed dose values, a
 #' logarithmic model instead of a log-logarithmic model should be used. Usually biological dose
@@ -128,6 +152,8 @@
 #'   grouping = peptide,
 #'   response = peptide_intensity_missing,
 #'   dose = concentration,
+#'   n_replicate_completeness = 2,
+#'   n_condition_completeness = 5,
 #'   retain_columns = c(protein, change_peptide)
 #' )
 #'
@@ -143,6 +169,10 @@ fit_drc_4p <- function(data,
                        filter = "post",
                        replicate_completeness = 0.7,
                        condition_completeness = 0.5,
+                       n_replicate_completeness = NULL,
+                       n_condition_completeness = NULL,
+                       complete_doses = NULL,
+                       anova_cutoff = 0.05,
                        correlation_cutoff = 0.8,
                        log_logarithmic = TRUE,
                        include_models = FALSE,
@@ -168,31 +198,42 @@ fit_drc_4p <- function(data,
   }
 
   if (filter != "none") {
-    n_conditions <- length(unique(dplyr::pull(data_prep, {{ dose }})))
-    n_replicates <- length(unique(dplyr::pull(data_prep, {{ sample }}))) / n_conditions
-    n_replicates_completeness <- floor(replicate_completeness * n_replicates)
-    n_conditions_completeness <- floor(condition_completeness * n_conditions)
+    if (!missing(complete_doses) & !is.null(complete_doses)) {
+      n_conditions <- length(complete_doses)
+      concentrations <- sort(complete_doses)
+    } else {
+      n_conditions <- length(unique(dplyr::pull(data_prep, {{ dose }})))
+      concentrations <- sort(unique(dplyr::pull(data_prep, {{ dose }})))
+    }
 
-    # perform anova on groups
-    anova <- data_prep %>%
-      dplyr::group_by({{ grouping }}, {{ dose }}) %>%
-      dplyr::mutate(
-        n = n_replicates,
-        mean_ratio = mean({{ response }}, na.rm = TRUE),
-        sd = sd({{ response }}, na.rm = TRUE)
-      ) %>%
-      dplyr::distinct({{ grouping }}, {{ dose }}, .data$mean_ratio, .data$sd, .data$n) %>%
-      tidyr::drop_na("mean_ratio", "sd") %>%
-      anova_protti({{ grouping }}, {{ dose }}, .data$mean_ratio, .data$sd, .data$n) %>%
-      dplyr::distinct({{ grouping }}, .data$pval) %>%
-      tidyr::drop_na("pval") %>% # remove NA pvalues before adjustment!
-      dplyr::mutate(anova_adj_pval = stats::p.adjust(.data$pval, method = "BH")) %>%
-      dplyr::rename(anova_pval = "pval")
+    if (!missing(condition_completeness) & !missing(n_condition_completeness) & !is.null(n_condition_completeness) & !is.null(condition_completeness)) {
+      warning("The condition_completeness argument will not be used in favor of using n_condition_completeness")
+    }
 
-    # extract elements that pass anova significant threshold
-    anova_filtered <- anova %>%
-      dplyr::filter(.data$anova_adj_pval <= 0.05) %>%
-      dplyr::pull({{ grouping }})
+    if (missing(n_condition_completeness) | is.null(n_condition_completeness)) {
+      lifecycle::deprecate_warn(
+        "0.8.0",
+        "fit_drc_4p(condition_completeness = )",
+        "fit_drc_4p(n_condition_completeness = )"
+      )
+
+      n_condition_completeness <- floor(condition_completeness * n_conditions)
+    }
+
+    if (!missing(replicate_completeness) & !missing(n_replicate_completeness) & !is.null(n_replicate_completeness) & !is.null(replicate_completeness)) {
+      warning("The replicate_completeness argument will not be used in favor of using n_replicate_completeness.")
+    }
+
+    if (missing(n_replicate_completeness) | is.null(n_replicate_completeness)) {
+      lifecycle::deprecate_warn(
+        "0.8.0",
+        "fit_drc_4p(replicate_completeness = )",
+        "fit_drc_4p(n_replicate_completeness = )"
+      )
+
+      n_replicates <- length(unique(dplyr::pull(data_prep, {{ sample }}))) / n_conditions
+      n_replicate_completeness <- floor(replicate_completeness * n_replicates)
+    }
 
     # filter for data completeness based on replicates and conditions
     # and based on the distribution of condition missingness
@@ -206,17 +247,16 @@ fit_drc_4p <- function(data,
     vector_add <- c(vector[1] + 1, vector[2] - 1)
     vector_add_rev <- rev(vector_add)
 
-    concentrations <- sort(unique(dplyr::pull(data_prep, {{ dose }})))
-
     filter_completeness <- data_prep %>%
       dplyr::group_by({{ grouping }}, {{ dose }}) %>%
       dplyr::mutate(enough_replicates = sum(!is.na({{ response }}))
-      >= n_replicates_completeness) %>%
+      >= n_replicate_completeness) %>%
+      dplyr::group_by({{ grouping }}, .data$enough_replicates) %>%
+      dplyr::mutate(n_condition_enough = n_distinct({{ dose }})) %>%
       dplyr::group_by({{ grouping }}) %>%
-      dplyr::mutate(enough_conditions = sum(.data$enough_replicates) /
-        n_replicates
-      >= n_conditions_completeness) %>%
-      dplyr::mutate(anova_significant = {{ grouping }} %in% anova_filtered) %>%
+      dplyr::mutate(enough_conditions = any(.data$n_condition_enough >= n_condition_completeness &
+        .data$enough_replicates)) %>%
+      dplyr::select(-"n_condition_enough") %>%
       dplyr::mutate(
         lower_vector = {{ dose }} %in% concentrations[1:vector[1]],
         lower_vector_rev = {{ dose }} %in% concentrations[1:vector_rev[1]],
@@ -281,45 +321,19 @@ fit_drc_4p <- function(data,
         list(unique(.data$n_vector_add_rev))[[1]][2]
       )$pval <= 0.1) %>%
       dplyr::ungroup() %>%
-      dplyr::mutate(
-        pval_vector = ifelse(is.na(.data$pval_vector),
-          TRUE,
-          .data$pval_vector
-        ),
-        pval_vector_rev = ifelse(is.na(.data$pval_vector_rev),
-          TRUE,
-          .data$pval_vector_rev
-        ),
-        pval_vector_add = ifelse(is.na(.data$pval_vector_add),
-          TRUE,
-          .data$pval_vector_add
-        ),
-        pval_vector_add_rev = ifelse(is.na(.data$pval_vector_add_rev),
-          TRUE,
-          .data$pval_vector_add_rev
-        )
-      ) %>%
-      dplyr::mutate(mean_vector = ifelse(is.na(.data$mean_vector),
-        0,
-        .data$mean_vector
-      )) %>%
-      dplyr::mutate(mean_vector_rev = ifelse(is.na(.data$mean_vector_rev),
-        0,
-        .data$mean_vector_rev
-      )) %>%
-      dplyr::mutate(mean_vector_add = ifelse(is.na(.data$mean_vector_add),
-        0,
-        .data$mean_vector_add
-      )) %>%
-      dplyr::mutate(mean_vector_add_rev = ifelse(is.na(.data$mean_vector_add_rev),
-        0,
-        .data$mean_vector_add_rev
-      )) %>%
+      tidyr::replace_na(list("pval_vector" = TRUE,
+                             "pval_vector_rev" = TRUE,
+                             "pval_vector_add" = TRUE,
+                             "pval_vector_add_rev" = TRUE,
+                             "mean_vector" = 0,
+                             "mean_vector_rev" = 0,
+                             "mean_vector_add" = 0,
+                             "mean_vector_add_rev" = 0)) %>%
       dplyr::group_by({{ grouping }}) %>%
-      dplyr::mutate(mean_vector = min(.data$mean_vector) == .data$mean_vector) %>%
-      dplyr::mutate(mean_vector_rev = min(.data$mean_vector_rev) == .data$mean_vector_rev) %>%
-      dplyr::mutate(mean_vector_add = min(.data$mean_vector_add) == .data$mean_vector_add) %>%
-      dplyr::mutate(mean_vector_add_rev = min(.data$mean_vector_add_rev) == .data$mean_vector_add_rev) %>%
+      dplyr::mutate(mean_vector = min(.data$mean_vector) == .data$mean_vector,
+                    mean_vector_rev = min(.data$mean_vector_rev) == .data$mean_vector_rev,
+                    mean_vector_add = min(.data$mean_vector_add) == .data$mean_vector_add,
+                    mean_vector_add_rev = min(.data$mean_vector_add_rev) == .data$mean_vector_add_rev) %>%
       dplyr::mutate(dose_MNAR = ifelse((all((.data$lower_vector & .data$enough_replicates == FALSE & .data$mean_vector) |
         (!.data$lower_vector & .data$enough_replicates == TRUE & !.data$mean_vector)) &
         .data$pval_vector) |
@@ -347,27 +361,42 @@ fit_drc_4p <- function(data,
       TRUE,
       FALSE
       )) %>%
-      dplyr::distinct({{ grouping }}, .data$enough_conditions, .data$anova_significant, .data$dose_MNAR) %>%
-      dplyr::mutate(passed_filter = (.data$enough_conditions == TRUE & .data$anova_significant == TRUE) |
-        .data$dose_MNAR == TRUE)
+      dplyr::distinct({{ grouping }}, .data$enough_conditions, .data$dose_MNAR)
 
     if (filter == "pre") {
       filtered_vector <- filter_completeness %>%
-        dplyr::filter(.data$passed_filter == TRUE) %>%
+        dplyr::filter(.data$enough_conditions) %>%
         dplyr::pull({{ grouping }})
 
       data_prep <- data_prep %>%
         filter({{ grouping }} %in% filtered_vector)
     }
+
+    # perform anova on groups
+    anova <- data_prep %>%
+      dplyr::group_by({{ grouping }}, {{ dose }}) %>%
+      dplyr::mutate(
+        n = sum(!is.na({{ response }})),
+        mean_ratio = mean({{ response }}, na.rm = TRUE),
+        sd = sd({{ response }}, na.rm = TRUE)
+      ) %>%
+      dplyr::distinct({{ grouping }}, {{ dose }}, .data$mean_ratio, .data$sd, .data$n) %>%
+      tidyr::drop_na("mean_ratio", "sd") %>%
+      anova_protti({{ grouping }}, {{ dose }}, .data$mean_ratio, .data$sd, .data$n) %>%
+      dplyr::distinct({{ grouping }}, .data$pval) %>%
+      tidyr::drop_na("pval") %>% # remove NA pvalues before adjustment!
+      dplyr::rename(anova_pval = "pval")
   }
   # prepare data
 
-  input <- data_prep %>%
+  input_prep <- data_prep %>%
     tidyr::drop_na({{ response }}) %>%
     dplyr::group_by({{ grouping }}) %>%
     dplyr::mutate(n_concentrations = dplyr::n_distinct(!!ensym(dose))) %>%
     dplyr::ungroup() %>%
-    split(dplyr::pull(., !!ensym(grouping))) %>%
+    split(dplyr::pull(., !!ensym(grouping)))
+
+  input <- input_prep %>%
     # make sure that there are enough data points to even fit a curve. This is not really filtering.
     purrr::keep(.p = ~ unique(.x$n_concentrations) > 4)
 
@@ -444,9 +473,21 @@ fit_drc_4p <- function(data,
       .f = ~ dplyr::mutate(.x, {{ grouping }} := .y)
     )
 
-  # Return empty data.frame if there are no correlations. This prevents parallel_fit_drc_4p from failing.
+  # Create data.frame if there are no correlations.
   if (nrow(correlation_output) == 0) {
-    return(data.frame())
+    p_value_correlation <- data.frame(pval = NA) %>% dplyr::bind_cols(filter_completeness %>%
+      dplyr::distinct({{ grouping }}))
+
+    correlation_output <- data.frame(correlation = NA) %>% dplyr::bind_cols(filter_completeness %>%
+      dplyr::distinct({{ grouping }}))
+
+    groups <- filter_completeness %>%
+      dplyr::pull({{ grouping }})
+
+    fit_objects <- setNames(
+      rep(list(list("coefficients" = c("hill:(Intercept)" = NA, "min_value:(Intercept)" = NA, "max_value:(Intercept)" = NA, "ec_50:(Intercept)" = NA))), length(groups)),
+      groups
+    )
   }
 
   # creating correlation output data frame
@@ -512,7 +553,7 @@ fit_drc_4p <- function(data,
       .f = ~ dplyr::mutate(.x, {{ grouping }} := .y)
     )
 
-  plot_points <- input[unique(dplyr::pull(line_fit, {{ grouping }}))] %>%
+  plot_points <- input_prep %>%
     purrr::map2_df(
       .y = names(.),
       .f = ~ dplyr::mutate(.x, {{ grouping }} := .y) %>%
@@ -521,7 +562,24 @@ fit_drc_4p <- function(data,
 
   # combining correlations with information for plot
 
+  no_fits <- as.data.frame(names(input_prep)[!names(input_prep) %in% pull(correlation_output, {{ grouping }})])
+  colnames(no_fits) <- rlang::as_name(rlang::enquo(grouping))
+
+  if (nrow(line_fit) == 0) {
+    line_fit <- as.data.frame(groups)
+    colnames(line_fit) <- rlang::as_name(rlang::enquo(grouping))
+
+    line_fit <- line_fit %>%
+      dplyr::mutate(
+        dose = NA,
+        Prediction = NA,
+        Lower = NA,
+        Upper = NA
+      )
+  }
+
   output <- correlation_output %>%
+    bind_rows(no_fits) %>%
     dplyr::left_join(line_fit, by = rlang::as_name(rlang::enquo(grouping))) %>%
     dplyr::group_by({{ grouping }}) %>%
     tidyr::nest(plot_curve = c("dose", "Prediction", "Lower", "Upper")) %>%
@@ -541,22 +599,24 @@ fit_drc_4p <- function(data,
     output <- output %>%
       dplyr::left_join(filter_completeness, by = rlang::as_name(rlang::enquo(grouping))) %>%
       dplyr::left_join(anova, by = rlang::as_name(rlang::enquo(grouping))) %>%
-      dplyr::mutate(passed_filter = .data$passed_filter &
+      dplyr::mutate(anova_adj_pval = stats::p.adjust(.data$anova_pval, method = "BH")) %>%
+      dplyr::mutate(anova_significant = ifelse(.data$anova_adj_pval > anova_cutoff | is.na(.data$anova_adj_pval),
+        FALSE,
+        TRUE
+      )) %>%
+      dplyr::mutate(passed_filter = (((.data$enough_conditions == TRUE & .data$anova_significant == TRUE) |
+        (.data$dose_MNAR == TRUE & .data$enough_conditions == TRUE)) &
         .data$correlation >= correlation_cutoff &
-        .data$min_model < .data$max_model) %>%
+        .data$min_model < .data$max_model) |
+        (.data$dose_MNAR & .data$enough_conditions)) %>%
       dplyr::group_by(.data$passed_filter) %>%
-      dplyr::mutate(score = ifelse(.data$passed_filter,
+      dplyr::mutate(score = ifelse(.data$passed_filter & .data$anova_significant & .data$correlation >= correlation_cutoff,
         (scale_protti(-log10(.data$anova_pval), method = "01") + scale_protti(.data$correlation, method = "01")) / 2,
         NA
       )) %>%
-      dplyr::ungroup()
+      dplyr::ungroup() %>%
+      dplyr::mutate(score = ifelse(is.na(.data$score) & .data$passed_filter == TRUE, 0, .data$score))
   }
-
-  if (filter == "pre") {
-    output <- output %>%
-      filter(.data$passed_filter)
-  }
-  # return result
 
   if (!missing(retain_columns)) {
     output <- data %>%